web_agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
datasets==3.5.1
dill==0.3.8
distro==1.9.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
jiter==0.9.0
multidict==6.4.3
multiprocess==0.70.16
numpy==2.2.5
openai==1.77.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
playwright==1.52.0
propcache==0.3.1
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
aiohttp-cors==0.7.0
